+++ src/libs/bin/top.rs
@@ -1,6 +1,6 @@
 use std::{
     collections::{BTreeMap, HashMap},
-    io::Write,
+    io::{IsTerminal, Write},
     rc::Rc,
 };
 
@@ -418,11 +418,21 @@ fn print(
     _env: &mut Environment,
     _ctx: &Expression,
 ) -> Result<Expression, RuntimeError> {
+    let is_tty = std::io::stdout().is_terminal();
     let mut stdout = std::io::stdout().lock();
     for x in args.iter() {
-        let _ = write!(&mut stdout, "{x} ");
+        let s = format!("{x} ");
+        if is_tty {
+            let _ = write!(&mut stdout, "{}", s.replace('\n', "\r\n"));
+        } else {
+            let _ = write!(&mut stdout, "{s}");
+        }
     }
-    let _ = writeln!(&mut stdout);
+    if is_tty {
+        let _ = write!(&mut stdout, "\r\n");
+    } else {
+        let _ = writeln!(&mut stdout);
+    }
     let _ = stdout.flush();
     Ok(Expression::None)
 }
@@ -431,9 +441,15 @@ fn println(
     _env: &mut Environment,
     _ctx: &Expression,
 ) -> Result<Expression, RuntimeError> {
+    let is_tty = std::io::stdout().is_terminal();
     let mut stdout = std::io::stdout().lock();
     for x in args.iter() {
-        let _ = writeln!(&mut stdout, "{x}");
+        let s = format!("{x}");
+        if is_tty {
+            let _ = write!(&mut stdout, "{}\r\n", s.replace('\n', "\r\n"));
+        } else {
+            let _ = writeln!(&mut stdout, "{s}");
+        }
     }
     let _ = stdout.flush();
     Ok(Expression::None)
@@ -454,12 +470,21 @@ fn eprint(
     _env: &mut Environment,
     _ctx: &Expression,
 ) -> Result<Expression, RuntimeError> {
+    let is_tty = std::io::stderr().is_terminal();
     let mut stderr = std::io::stderr().lock();
     for (i, x) in args.iter().enumerate() {
-        if i < args.len() - 1 {
-            let _ = write!(&mut stderr, "\x1b[38;5;9m{x} \x1b[m\x1b[0m");
+        let s = format!("\x1b[38;5;9m{x}\x1b[m\x1b[0m");
+        if is_tty {
+            let s = s.replace('\n', "\r\n");
+            if i < args.len() - 1 {
+                let _ = write!(&mut stderr, "{s} ");
+            } else {
+                let _ = writeln!(&mut stderr, "{s}");
+            }
+        } else if i < args.len() - 1 {
+            let _ = write!(&mut stderr, "{s} ");
         } else {
-            let _ = writeln!(&mut stderr, "\x1b[38;5;9m{x}\x1b[m\x1b[0m");
+            let _ = writeln!(&mut stderr, "{s}");
         }
     }
     let _ = stderr.flush();
@@ -470,9 +495,15 @@ fn eprintln(
     _env: &mut Environment,
     _ctx: &Expression,
 ) -> Result<Expression, RuntimeError> {
+    let is_tty = std::io::stderr().is_terminal();
     let mut stderr = std::io::stderr().lock();
     for x in args.iter() {
-        let _ = writeln!(&mut stderr, "\x1b[38;5;9m{x}\x1b[m\x1b[0m");
+        let s = format!("\x1b[38;5;9m{x}\x1b[m\x1b[0m");
+        if is_tty {
+            let _ = writeln!(&mut stderr, "{}", s.replace('\n', "\r\n"));
+        } else {
+            let _ = writeln!(&mut stderr, "{s}");
+        }
     }
     let _ = stderr.flush();
     Ok(Expression::None)